Optimize NMS part 2
